Move soo_buffer to containers
